requirements.txt
fastapi==0.87.0
uvicorn==0.19.0
gunicorn==20.1.0
httpx==0.23.1
jupyter==1.0.0
starlette==0.21.0
pydantic==1.10.2
orjson==3.8.1

pandas==1.5.2
rich==12.6.0
rectools==0.2.0
numpy==1.23.5
joblib==1.2.0
scipy==1.8.1
pyyaml

click